fix: remove build section for docker compose
docker compose will ignore any build section in docker-compose.yml if
image section is specified, unlike the deprecated docker-compose.
And the build section is not needed anyway.

The docker-compose.test.yml will very likely not work with docker
compose because of image and build section specified together.
That PR does not remove the wrong image section in it.

Signed-off-by: Jérôme Benoit <jerome.benoit@piment-noir.org>